Refactoring the communication related functions into its own file
