+++ lib/context_adapter_helper.js
@@ -25,10 +25,7 @@
 
 var caLogger = require('logops'),
     caConfig = require('./context_adapter_configuration'),
-    caErrors = require('./context_adapter_error'),
-    caPackage = require('../package.json'),
-    request = require('request'),
-    revalidator = require('revalidator');
+    caPackage = require('../package.json');
 
 /**
  * Generates a 32 bit integer hash code
@@ -145,66 +142,6 @@ function getVersion() {
 }
 
 /**
- * Sends a queryContext request to the configured Context Broker
- * @param {string} service Service header to be used in the request
- * @param {string} servicePath Service Path header to be used in the request
- * @param {object} payload Payload to be included in the request
- * @param {function} callback Callback to be called informing of any errors or returning
- *  the obtained response
- */
-function sendQueryContext(service, servicePath, payload, callback) {
-  var requestOptions = {
-    method: 'POST',
-    uri: 'http://' + caConfig.CB_HOST + ':' + caConfig.CB_PORT + caConfig.CB_PATH + '/queryContext',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Fiware-Service': service || caConfig.SERVICE,
-      'Fiware-ServicePath': servicePath || caConfig.SERVICE_PATH
-    },
-    json: true,
-    body: payload
-  };
-  request(requestOptions, callback);
-}
-
-/**
- * Sends a updateContext request to the configured Context Broker
- * @param {string} service Service header to be used in the request
- * @param {string} servicePath Service Path header to be used in the request
- * @param {object} payload Payload to be included in the request
- * @param {function} callback Callback to be called informing of any errors or returning
- *  the obtained response
- */
-function sendUpdateContext(service, servicePath, payload, callback) {
-  var context = {
-      corr: getUnicaCorrelator(),
-      trans: getTransactionId(),
-    op: getOperationType()
-  };
-
-  caLogger.debug(
-    context,
-    'Sending updateContext to CB at ' + 'http://' + caConfig.CB_HOST + ':' + caConfig.CB_PORT + caConfig.CB_PATH +
-    '/updateContext' + ' with payload: ' + JSON.stringify(payload)
-  );
-
-  var requestOptions = {
-    method: 'POST',
-    uri: 'http://' + caConfig.CB_HOST + ':' + caConfig.CB_PORT + caConfig.CB_PATH + '/updateContext',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Fiware-Service': service || caConfig.SERVICE,
-      'Fiware-ServicePath': servicePath || caConfig.SERVICE_PATH
-    },
-    json: true,
-    body: payload
-  };
-  request(requestOptions, callback);
-}
-
-/**
  * Indicates if certain attribute is included in an attributes array
  * @param {Array} attributes The attributes array
  * @param {string} attributeName The attribute name to find
@@ -285,589 +222,10 @@ function stringifyRevalidatorErrors(errors) {
 }
 
 /**
- * Checks that an updateContext request sent to the Context Adapter is well-formed
- * @param {object} request The request sent to the Context Adapter
- * @param {function} callback Callback to be called with the result of the check
- * @return {*} boolean True if the updateContext request is valid. False otherwise
- */
-function checkUpdateContextRequest(request, callback) {
-  var updateContextJSONSchema = {
-    properties: {
-      payload: {
-        id: 'payload',
-        type: 'object',
-        required: true,
-        properties: {
-          contextElements: {
-            id: 'contextElements',
-            type: 'array',
-            required: true,
-            minItems: 0,
-            maxItems: 2,
-            items: {
-              id: '0',
-              type: 'object',
-              required: true,
-              properties: {
-                id: {
-                  id: 'id',
-                  type: 'string',
-                  required: true
-                },
-                type: {
-                  id: 'type',
-                  type: 'string',
-                  required: true
-                },
-                isPattern: {
-                  id: 'isPattern',
-                  type: 'string',
-                  required: true,
-                  pattern: 'false'
-                },
-                attributes: {
-                  id: 'attributes',
-                  type: 'array',
-                  required: true,
-                  minItems: 4,
-                  maxItems: 6,
-                  conform: function(attributes) {
-                    return isAttributeIncluded(attributes,
-                        caConfig.BUTTON_ENTITY.CA_INTERACTION_TYPE_ATTR_NAME) &&
-                      isAttributeIncluded(attributes,
-                        caConfig.BUTTON_ENTITY.CA_SERVICE_ID_ATTR_NAME) &&
-                      isAttributeIncluded(attributes,
-                        caConfig.BUTTON_ENTITY.CA_OPERATION_ACTION_ATTR_NAME) &&
-                      isAttributeIncluded(attributes,
-                        caConfig.BUTTON_ENTITY.CA_OPERATION_EXTRA_ATTR_NAME) &&
-                      (getAttributeValue(attributes, caConfig.BUTTON_ENTITY.CA_INTERACTION_TYPE_ATTR_NAME) ===
-                        caConfig.BUTTON_ENTITY.INTERACTION_TYPES.SYNCHRONOUS ||
-                          isAttributeIncluded(attributes, caConfig.BUTTON_ENTITY.CA_OPERATION_STATUS_ATTR_NAME)) &&
-                      isValidAttributeValues(request.contextAdapter.context, caConfig.BUTTON_ENTITY.TYPE, attributes);
-                  },
-                  message: 'lacks some mandatory attributes or invalid values are included'
-                }
-              }
-            }
-          },
-          updateAction: {
-            id: 'updateAction',
-            type: 'string',
-            required: true,
-            pattern: 'UPDATE'
-          }
-        }
-      }
-    }
-  };
-
-  var validation = revalidator.validate(request, updateContextJSONSchema);
-  if (validation.valid) {
-    return process.nextTick(callback);
-  } else {
-    var err = new caErrors.BadPayload(
-      request.payload
-    );
-    err.message += ' (errors: ' + stringifyRevalidatorErrors(validation.errors) + ')';
-    return process.nextTick(callback.bind(null, err));
-  }
-}
-
-/**
- * Processes and returns the requested operation information
- * @param {object} request The updateContext request received
- * @param {function} callback Callback to be called with the result of the operation
- */
-function getOperationDescriptor(request, callback) {
-  checkUpdateContextRequest(request, function(err) {
-    if (err) {
-      return process.nextTick(callback.bind(null, err));
-    } else {
-      var operationDescriptor = {
-        buttonId: request.payload.contextElements[0].id
-      };
-      request.payload.contextElements[0].attributes.forEach(function(attribute) {
-        switch (attribute.name) {
-          case caConfig.BUTTON_ENTITY.CA_INTERACTION_TYPE_ATTR_NAME:
-            operationDescriptor.interactionType = attribute.value;
-            break;
-          case caConfig.BUTTON_ENTITY.CA_SERVICE_ID_ATTR_NAME:
-            operationDescriptor.serviceId = attribute.value;
-            break;
-          case caConfig.BUTTON_ENTITY.CA_OPERATION_ACTION_ATTR_NAME:
-            operationDescriptor.action = attribute.value;
-            break;
-          case caConfig.BUTTON_ENTITY.CA_OPERATION_EXTRA_ATTR_NAME:
-            operationDescriptor.extra = attribute.value;
-            break;
-        }
-      });
-      request.contextAdapter.operationDescriptor = operationDescriptor;
-      return process.nextTick(callback.bind(null, null, operationDescriptor));
-    }
-  });
-}
-
-/**
- * Checks that an queryContext response returned by the Context Adapter to get
- *  service information is well-formed
- * @param {object} context The context for logging
- * @param {object} response The response received from the Context Adapter
- * @param {string} serviceId The identifier of the service whose information was retrieved
- * @param {function} callback Callback to be called with the result of the check
- * @return {*} True if the queryContextResponse is valid. False otherwise.
- */
-function checkQueryContextResponse(context, response, serviceId, callback) {
-  var queryContextJSONSchema = {
-    properties: {
-      body: {
-        id: 'body',
-        type: 'object',
-        required: true,
-        properties: {
-          contextResponses: {
-            id: 'contextResponses',
-            type: 'array',
-            required: true,
-            minItems: 0,
-            maxItems: 2,
-            items: {
-              id: '0',
-              type: 'object',
-              required: true,
-              properties: {
-                contextElement: {
-                  id: 'contextElement',
-                  type: 'object',
-                  required: true,
-                  properties: {
-                    id: {
-                      id: 'id',
-                      type: 'string',
-                      required: true
-                    },
-                    type: {
-                      id: 'type',
-                      type: 'string',
-                      required: true
-                    },
-                    isPattern: {
-                      id: 'isPattern',
-                      type: 'string',
-                      required: true,
-                      pattern: 'false'
-                    },
-                    attributes: {
-                      id: 'attributes',
-                      type: 'array',
-                      required: true,
-                      minItems: 6,
-                      maxItems: 8,
-                      conform: function(attributes) {
-                        return isAttributeIncluded(
-                          attributes,
-                          caConfig.SERVICE_ENTITY.PROVIDER_ATTR_NAME) &&
-                        isAttributeIncluded(
-                            attributes,
-                            caConfig.SERVICE_ENTITY.INTERACTION_TYPE_ATTR_NAME) &&
-                          isAttributeIncluded(
-                            attributes,
-                            caConfig.SERVICE_ENTITY.ENDPOINT_ATTR_NAME) &&
-                          isAttributeIncluded(
-                            attributes,
-                            caConfig.SERVICE_ENTITY.METHOD_ATTR_NAME) &&
-                          isAttributeIncluded(
-                            attributes,
-                            caConfig.SERVICE_ENTITY.AUTHENTICATION_ATTR_NAME) &&
-                          isAttributeIncluded(
-                            attributes,
-                            caConfig.SERVICE_ENTITY.MAPPING_ATTR_NAME) &&
-                          isAttributeIncluded(
-                            attributes,
-                            caConfig.SERVICE_ENTITY.TIMEOUT_ATTR_NAME) &&
-                          isValidAttributeValues(context, caConfig.SERVICE_ENTITY.TYPE, attributes);
-                      },
-                      message: 'lacks some mandatory attributes or invalid values are included'
-                    }
-                  }
-                },
-                statusCode: {
-                  id: 'statusCode',
-                  type: 'object',
-                  required: true,
-                  properties: {
-                    code: {
-                      id: 'code',
-                      type: 'string',
-                      required: true,
-                      pattern: '200'
-                    },
-                    reasonPhrase: {
-                      id: 'code',
-                      type: 'string',
-                      required: true,
-                      pattern: 'OK'
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  };
-
-  var validation = revalidator.validate(response, queryContextJSONSchema);
-  if (validation.valid) {
-    return process.nextTick(callback);
-  } else {
-    var err = new caErrors.BadService(serviceId);
-    err.message += ' (errors: ' + stringifyRevalidatorErrors(validation.errors) + ')';
-    return process.nextTick(callback.bind(null, err));
-  }
-}
-
-/**
- * Returns a service descriptor based on a service unique identifier
- * @param {object} request The received updateContext request
- * @param {object} operationDescriptor The operation descriptor associated to the request
- * @param {function} callback Callback to be called informing of any errors or returning
- *  the required service descriptor
- */
-function getServiceDescriptor(request, operationDescriptor, callback) {
-  var payload = {
-    entities: [
-      {
-        id: operationDescriptor.serviceId,
-        type: caConfig.SERVICE_ENTITY.TYPE,
-        isPattern: false
-      }
-    ],
-    attributes: [
-      caConfig.SERVICE_ENTITY.PROVIDER_ATTR_NAME,
-      caConfig.SERVICE_ENTITY.INTERACTION_TYPE_ATTR_NAME,
-      caConfig.SERVICE_ENTITY.ENDPOINT_ATTR_NAME,
-      caConfig.SERVICE_ENTITY.METHOD_ATTR_NAME,
-      caConfig.SERVICE_ENTITY.AUTHENTICATION_ATTR_NAME,
-      caConfig.SERVICE_ENTITY.MAPPING_ATTR_NAME,
-      caConfig.SERVICE_ENTITY.TIMEOUT_ATTR_NAME
-    ]
-  };
-
-  caLogger.debug(
-    request.contextAdapter.context,
-    'Sending queryContext to the Context Broker to get the service entity with payload: ' +
-    JSON.stringify(payload)
-  );
-
-  sendQueryContext(
-    request.headers['fiware-service'],
-    request.headers['fiware-servicepath'],
-    payload,
-    function(err, response, body) {
-      caLogger.debug(
-        request.contextAdapter.context,
-        'Response received from the Context Broker when querying for the service entity: ' +
-          'err: ' + err + ', payload: ' + JSON.stringify(body)
-      );
-
-      if (err) {
-        return process.nextTick(callback.bind(null, err));
-      }
-
-      checkQueryContextResponse(request.contextAdapter.context, response, operationDescriptor.serviceId, function(err) {
-        if (err) {
-          return process.nextTick(callback.bind(null, err));
-        } else {
-          var serviceDescription = {
-            id: operationDescriptor.serviceId
-          };
-          body.contextResponses[0].contextElement.attributes.forEach(function(attribute) {
-            switch (attribute.name) {
-              case caConfig.SERVICE_ENTITY.PROVIDER_ATTR_NAME:
-                serviceDescription.provider = attribute.value;
-                break;
-              case caConfig.SERVICE_ENTITY.INTERACTION_TYPE_ATTR_NAME:
-                serviceDescription.interactionType = attribute.value;
-                break;
-              case caConfig.SERVICE_ENTITY.ENDPOINT_ATTR_NAME:
-                serviceDescription.endpoint = attribute.value;
-                break;
-              case caConfig.SERVICE_ENTITY.METHOD_ATTR_NAME:
-                serviceDescription.method = attribute.value;
-                break;
-              case caConfig.SERVICE_ENTITY.AUTHENTICATION_ATTR_NAME:
-                serviceDescription.authentication = attribute.value;
-                break;
-              case caConfig.SERVICE_ENTITY.MAPPING_ATTR_NAME:
-                serviceDescription.mapping = attribute.value;
-                break;
-              case caConfig.SERVICE_ENTITY.TIMEOUT_ATTR_NAME:
-                serviceDescription.timeout = attribute.value;
-                break;
-            }
-          });
-          request.contextAdapter.serviceDescriptor = serviceDescription;
-          return process.nextTick(callback.bind(null, null, serviceDescription));
-        }
-      });
-    }
-  );
-}
-
-/**
- * Processes and returns the payload to be sent in the request to a Third Party
- *  based on certain service and operation descriptors
- * @param {object} serviceDescriptor The service descriptor
- * @param {object} operationDescriptor The operation descriptor
- * @return {object} The payload to be sent in the request to a Third Party
- */
-function getThirdPartyPayload(serviceDescriptor, operationDescriptor) {
-  var payload = {
-    button: operationDescriptor.buttonId,
-    action: operationDescriptor.action,
-    extra: operationDescriptor.extra
-  };
-  if (serviceDescriptor.interactionType ===
-      caConfig.SERVICE_ENTITY.INTERACTION_TYPES.ASYNCHRONOUS) {
-    payload.callback = 'http://' + caConfig.CA_HOST + ':' + caConfig.CA_PORT +
-      caConfig.CA_PATH + caConfig.CA_CALLBACK_PATH;
-  }
-  return payload;
-}
-
-/**
- * Generates the result in case of error or from the response obtained from the third party
- * @param {object} err The error, if any
- * @param {object} data The data returned by the Third Party as the operation request result
- * @return {object} The generated operation result
- */
-function getResult(err, data) {
-  if (err) {
-    // TODO For the time being we return '0,<error code>,<details>'
-    return '0,' + err.code + ',' + err.message;
-  }
-
-  var detailsObj;
-  var result = '';
-  if (data) {
-    if (data.body) {
-      detailsObj = data.body.details;
-    } else if (data.payload) {
-      detailsObj = data.payload.details;
-    } else {
-      return '0,' + 'XXX' + ',' + 'Unknown data received from the Third Party as the request result';
-    }
-    for (var key in detailsObj) {
-      if (detailsObj.hasOwnProperty(key)) {
-        result += encodeURIComponent(key + '-' + detailsObj[key] + ';');
-      }
-    }
-    return '1,' + result;
-  } else {
-    return '0,XXX,Unknown data received from the Third Party as the request result';
-  }
-}
-
-/**
- * Sends a synchronous request to the third party
- * @param {object} updateContextRequest The updateContext request received
- * @param {function} callback Callback to be called informing of any errors or returning
- *  the obtained response
- */
-function sendThirdPartyRequest(updateContextRequest, callback) {
-  var payload = getThirdPartyPayload(updateContextRequest.contextAdapter.serviceDescriptor,
-    updateContextRequest.contextAdapter.operationDescriptor);
-
-  caLogger.debug(
-    updateContextRequest.contextAdapter.context,
-    'Sending ' + updateContextRequest.contextAdapter.serviceDescriptor.method + ' request to third party at: ' +
-      updateContextRequest.contextAdapter.serviceDescriptor.endpoint + ' with headers: ' +
-      'Fiware-Service: ' + updateContextRequest.headers['fiware-service'] + ' and ' +
-      'Fiware-ServicePath: ' + updateContextRequest.headers['fiware-servicepath'] +
-      ' and payload: ' + JSON.stringify(payload)
-  );
-
-  var requestOptions = {
-    method: updateContextRequest.contextAdapter.serviceDescriptor.method,
-    uri: updateContextRequest.contextAdapter.serviceDescriptor.endpoint,
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Fiware-Service': updateContextRequest.headers['fiware-service'],
-      'Fiware-ServicePath': updateContextRequest.headers['fiware-servicepath']
-    },
-    json: true,
-    body: payload
-  };
-  request(requestOptions, callback);
-}
-
-/**
- * Sends a operation completed notification to the Context Broker
- * @param {object} request The updateContext request received by the Context Adapter if available
- * @param {object} data The data received from the Third Party as the operation request result
- *  an updateContext request, null if the request is an update request sent by a third party.
- * @param {function} callback Callback to be called informing of any errors or returning
- *  the result of the notification
- */
-function notifyOperationCompleted(request, data, callback) {
-  var contextElements = [
-    {
-      id: request ? request.contextAdapter.operationDescriptor.buttonId :
-        data.payload.buttonId,
-      type: caConfig.BUTTON_ENTITY.TYPE,
-      isPattern: false,
-      attributes: [
-        {
-          name: caConfig.BUTTON_ENTITY.OPERATION_STATUS_ATTR_NAME,
-          type: caConfig.ATTRIBUTE_TYPE.STRING,
-          value: caConfig.BUTTON_ENTITY.OPERATION_STATUS.COMPLETED
-        },
-        {
-          name: caConfig.BUTTON_ENTITY.OPERATION_RESULT_ATTR_NAME,
-          type: caConfig.ATTRIBUTE_TYPE.STRING,
-          value: getResult(null, data)
-        }
-      ]
-    }
-  ];
-
-  if (request && request.contextAdapter.operationDescriptor.interactionType !==
-    caConfig.BUTTON_ENTITY.INTERACTION_TYPES.ASYNCHRONOUS) {
-    contextElements[0].attributes.push({
-      name: caConfig.BUTTON_ENTITY.IOTA_OPERATION_RESULT_ATTR_NAME,
-      type: caConfig.ATTRIBUTE_TYPE.STRING,
-      value: getResult(null, data)
-    });
-  }
-
-  var payload = {
-    contextElements: contextElements,
-    updateAction: 'UPDATE'
-  };
-
-  caLogger.debug(
-    request ? request.contextAdapter.context : data.contextAdapter.context,
-    'Notifying the Context Broker that the request has been completed with headers ' +
-    'Fiware-Service: ' + (request ? request.headers['fiware-service'] : data.headers['fiware-service']) + ' and ' +
-    'Fiware-ServicePath: ' + (request ? request.headers['fiware-servicepath'] : data.headers['fiware-servicepath']) +
-    ' and payload: ' + JSON.stringify(payload)
-  );
-
-  sendUpdateContext(
-    request ? request.headers['fiware-service'] : data.headers['fiware-service'],
-    request ? request.headers['fiware-servicepath'] : data.headers['fiware-servicepath'],
-    payload,
-    callback
-  );
-}
-
-/**
- * Sends a operation closed notification to the Context Broker
- * @param {object} err The error which caused the request to be cancelled and closed
- * @param {request} request The received updateContext request
- * @param {function} callback Callback to be called informing of any errors or returning
- *  the result of the notification
- */
-function notifyOperationClosed(err, request, callback) {
-  var payload = {
-    contextElements: [
-      {
-        id: request.contextAdapter.operationDescriptor.buttonId,
-        type: caConfig.BUTTON_ENTITY.TYPE,
-        isPattern: false,
-        attributes: [
-          {
-            name: caConfig.BUTTON_ENTITY.OPERATION_STATUS_ATTR_NAME,
-            type: caConfig.ATTRIBUTE_TYPE.STRING,
-            value: caConfig.BUTTON_ENTITY.OPERATION_STATUS.CLOSED
-          },
-          {
-            name: caConfig.BUTTON_ENTITY.IOTA_OPERATION_STATUS_ATTR_NAME,
-            type: caConfig.ATTRIBUTE_TYPE.STRING,
-            value: caConfig.BUTTON_ENTITY.OPERATION_STATUS.CLOSED
-          },
-          {
-            name: caConfig.BUTTON_ENTITY.OPERATION_RESULT_ATTR_NAME,
-            type: caConfig.ATTRIBUTE_TYPE.STRING,
-            value: getResult(err)
-          },
-          {
-            name: caConfig.BUTTON_ENTITY.IOTA_OPERATION_RESULT_ATTR_NAME,
-            type: caConfig.ATTRIBUTE_TYPE.STRING,
-            value: getResult(err)
-          }
-        ]
-      }
-    ],
-    updateAction: 'UPDATE'
-  };
-
-  caLogger.debug(
-    request.contextAdapter.context,
-    'Notifying the Context Broker that the request has been closed with headers : ' +
-    'Fiware-Service: ' + request.headers['fiware-service'] + ' and ' +
-    'Fiware-ServicePath: ' + request.headers['fiware-servicepath'] +
-    ' and payload: ' + JSON.stringify(payload)
-  );
-
-  sendUpdateContext(
-    request.headers['fiware-service'],
-    request.headers['fiware-servicepath'],
-    payload,
-    callback
-  );
-}
-
-/**
- * Returns the NGSI response for an updateContext request based on the
- *  request itself and on the existence of any error during its processing
- * @param {object} err The error when processing the request, if any
- * @param {object} request The updateContext request
- * @return {object} The composed response
- */
-function getNGSIResponse(err, request) {
-  var response = {
-    contextResponses: []
-  };
-  var statusCode;
-  if (err) {
-    statusCode = {
-      code: '400',
-      reasonPhrase: encodeURIComponent(err.code + ' - ' + err.message)
-    };
-  } else {
-    statusCode = {
-      code: '200',
-      reasonPhrase: 'OK'
-    };
-  }
-  if (request.payload && request.payload.contextElements) {
-    request.payload.contextElements.forEach(function(contextElement) {
-      var entry = {
-        contextElement: contextElement,
-        statusCode: statusCode
-      };
-      response.contextResponses.push(entry);
-    });
-  } else {
-    response.contextResponses.push({
-      statusCode: statusCode
-    });
-  }
-  return response;
-}
-
-/**
- * * Properties and functions exported by the module
+ * Properties and functions exported by the module
  * @type {{getUnicaCorrelator: getUnicaCorrelator, getTransactionId: getTransactionId,
  * getOperationType: getOperationType, getVersion: getVersion, getAttributeValue: getAttributeValue,
- * setAttribute: setAttribute, removeAttribute: removeAttribute, getNGSIResponse: getNGSIResponse,
- * getServiceDescriptor: getServiceDescriptor, getOperationDescriptor: getOperationDescriptor,
- * sendThirdPartyRequest: sendThirdPartyRequest, notifyOperationCompleted: notifyOperationCompleted,
- * notifyOperationClosed: notifyOperationClosed, getResult: getResult}}
+ * setAttribute: setAttribute, removeAttribute: removeAttribute}}
  */
 module.exports = {
   getUnicaCorrelator: getUnicaCorrelator,
@@ -877,11 +235,7 @@ module.exports = {
   getAttributeValue: getAttributeValue,
   setAttribute: setAttribute,
   removeAttribute: removeAttribute,
-  getNGSIResponse: getNGSIResponse,
-  getServiceDescriptor: getServiceDescriptor,
-  getOperationDescriptor: getOperationDescriptor,
-  sendThirdPartyRequest: sendThirdPartyRequest,
-  notifyOperationCompleted: notifyOperationCompleted,
-  notifyOperationClosed: notifyOperationClosed,
-  getResult: getResult
+  isAttributeIncluded: isAttributeIncluded,
+  isValidAttributeValues: isValidAttributeValues,
+  stringifyRevalidatorErrors: stringifyRevalidatorErrors
 };
