+++ lib/context_adapter_server.js
@@ -25,6 +25,7 @@
 
 var caLogger = require('logops'),
     caConfig = require('./context_adapter_configuration'),
+    caComm = require('./context_adapter_communication'),
     caErrors = require('./context_adapter_error'),
     caHelper = require('./context_adapter_helper.js'),
     hapi = require('hapi'),
@@ -106,7 +107,7 @@ function respond2UpdateContext(request, reply, operationDescriptor, callback) {
     JSON.stringify(operationDescriptor)
   );
   // Respond to the received updateContext request
-  var ngsiResponse = caHelper.getNGSIResponse(null, request);
+  var ngsiResponse = caComm.getNGSIResponse(null, request);
   reply(ngsiResponse);
   callback(null, operationDescriptor);
 }
@@ -126,7 +127,7 @@ function updateContextErrorHandler(request, reply, err) {
 
   if (err instanceof caErrors.BadPayload) {
     // Respond to the received updateContext request
-    var ngsiResponse = caHelper.getNGSIResponse(err, request);
+    var ngsiResponse = caComm.getNGSIResponse(err, request);
     caLogger.debug(
       request.contextAdapter.context,
       'Responding to the Context Adapter with payload: ' +
@@ -135,7 +136,7 @@ function updateContextErrorHandler(request, reply, err) {
     reply(ngsiResponse);
   } else {
     // Notify the operation as closed
-    caHelper.notifyOperationClosed(
+    caComm.notifyOperationClosed(
       err,
       request,
       function onNotifyOperationClosed() {
@@ -160,7 +161,7 @@ function onThirdPartyResponse(request, err, response) {
         err.code + ' - ' + err.message
     );
     // Notify the operation as closed
-    caHelper.notifyOperationClosed(
+    caComm.notifyOperationClosed(
       err,
       request,
       function onNotifyOperationClosed() {
@@ -178,7 +179,7 @@ function onThirdPartyResponse(request, err, response) {
     if (request.contextAdapter.serviceDescriptor.interactionType ===
       caConfig.SERVICE_ENTITY.INTERACTION_TYPES.SYNCHRONOUS) {
       // Notify the operation as completed
-      caHelper.notifyOperationCompleted(
+      caComm.notifyOperationCompleted(
         request,
         response,
         function onNotifyOperationCompleted(err, response, body) {
@@ -214,7 +215,7 @@ function processOperationRequest(request, serviceDescriptor) {
   );
 
   // Send a request to the Third Party
-  caHelper.sendThirdPartyRequest(request, onThirdPartyResponse.bind(null, request));
+  caComm.sendThirdPartyRequest(request, onThirdPartyResponse.bind(null, request));
 }
 
 /**
@@ -247,9 +248,9 @@ function updateContextHandler(request, reply) {
   );
 
   async.waterfall([
-    caHelper.getOperationDescriptor.bind(null, request),
+    caComm.getOperationDescriptor.bind(null, request),
     respond2UpdateContext.bind(null, request, reply),
-    caHelper.getServiceDescriptor.bind(null, request),
+    caComm.getServiceDescriptor.bind(null, request),
     processOperationRequest.bind(null, request)
   ], updateContextErrorHandler.bind(null, request, reply));
 }
@@ -272,7 +273,7 @@ function thirdPartyUpdateHandler(request, reply) {
   );
 
   // Notify the operation as completed
-  caHelper.notifyOperationCompleted(
+  caComm.notifyOperationCompleted(
     null,
     request,
     function onNotifyOperationCompleted(err, response, body) {
